Moved OmdbMediaResponse into models
